Added margin call tests

+++ contracts/interfaces/Invest/ICToken.sol
@@ -22,7 +22,7 @@ interface ICToken {
 
     function exchangeRateCurrent() external returns (uint256);
 
-    function exchangeRateStored() external view returns (uint);
+    function exchangeRateStored() external view returns (uint256);
 
     function supplyRatePerBlock() external returns (uint256);
